+++ src/main/java/com/airesnor/wuxiacraft/capabilities/CultivationStorage.java
@@ -16,8 +16,8 @@ public class CultivationStorage implements Capability.IStorage<ICultivation> {
 		NBTTagCompound tag = new NBTTagCompound();
 		tag.setString("level", instance.getCurrentLevel().name());
 		tag.setInteger("subLevel", instance.getCurrentSubLevel());
-		tag.setInteger("progress", (int) instance.getCurrentProgress());
-		tag.setInteger("energy", (int) instance.getEnergy());
+		tag.setDouble("progress", instance.getCurrentProgress());
+		tag.setFloat("energy", instance.getEnergy());
 		tag.setInteger("pelletCD", instance.getPillCooldown());
 		tag.setBoolean("suppress", instance.getSuppress());
 		return tag;
@@ -28,8 +28,8 @@ public class CultivationStorage implements Capability.IStorage<ICultivation> {
 		NBTTagCompound tag = (NBTTagCompound) nbt;
 		instance.setCurrentLevel(CultivationLevel.valueOf(tag.getString("level")));
 		instance.setCurrentSubLevel(tag.getInteger("subLevel"));
-		instance.addProgress(tag.getInteger("progress"));
-		instance.addEnergy(tag.getInteger("energy"));
+		instance.addProgress(tag.getDouble("progress"));
+		instance.addEnergy(tag.getFloat("energy"));
 		instance.setPillCooldown(tag.getInteger("pelletCD"));
 		instance.setSuppress(tag.getBoolean("suppress"));
 	}
