+++ src/main/java/com/airesnor/wuxiacraft/networking/ProgressMessage.java
@@ -6,14 +6,14 @@ import net.minecraftforge.fml.common.network.simpleimpl.IMessage;
 public class ProgressMessage implements IMessage {
 
 	public int op;
-	public float amount;
+	public double amount;
 
 	public ProgressMessage() {
 		this.op = 0;
 		this.amount = 0;
 	}
 
-	public ProgressMessage(int op, float amount) {
+	public ProgressMessage(int op, double amount) {
 		this.op = op;
 		this.amount = amount;
 	}
@@ -21,12 +21,12 @@ public class ProgressMessage implements IMessage {
 	@Override
 	public void fromBytes(ByteBuf buf) {
 		this.op = buf.readInt();
-		this.amount = buf.readFloat();
+		this.amount = buf.readDouble();
 	}
 
 	@Override
 	public void toBytes(ByteBuf buf) {
 		buf.writeInt(this.op);
-		buf.writeFloat(this.amount);
+		buf.writeDouble(this.amount);
 	}
 }
